Migrate e2e-encryption tests to OpenShift Tests Extension framework
This commit migrates five test suites (e2e-encryption, e2e-encryption-kms,
e2e-encryption-perf, e2e-encryption-rotation, e2e-oidc) from standalone
test binaries to the OpenShift Tests Extension (OTE) framework using Ginkgo v2.

Key changes:
- Migrate tests to Ginkgo v2 with OTE integration
- Extract test logic into dedicated helper files for better organization
- Remove obsolete main_test.go files in favor of Ginkgo suite setup
- Configure appropriate test suite parallelism and qualifiers
- Add klog stderr redirection to fix JSON output corruption

Critical fix for klog output corruption:
- Add klog.SetOutput(os.Stderr) in test binary's main() to prevent klog
  warnings from corrupting JSON output during 'openshift-tests list' commands
- This fixes "invalid character 'W'" parsing errors in CI that caused PR #839
  to be reverted in PR #857
- Ensures stdout contains only clean JSON for CI consumption while klog
  warnings go to stderr

Test organization:
- operator/parallel: Non-serial operator, templates, and tokens tests (parallelism: 4)
- operator/serial: Serial operator, OIDC, templates, and tokens tests
- operator/disruptive: Tests marked as [Disruptive]
- operator-encryption/serial: Basic encryption tests
- operator-encryption-rotation/serial: Encryption key rotation tests
- operator-encryption-perf/serial: Encryption performance tests
- operator-encryption-kms/serial: KMS encryption tests

Dependencies:
- Use standard OTE version v0.0.0-20251205182537-ff5553e56f33
- Updated klog dependency to v2.140.0
- No fork dependencies

See: https://github.com/openshift/cluster-authentication-operator/pull/857 (revert)
     https://github.com/openshift/cluster-authentication-operator/pull/859 (this PR)
